CORTX-31563 Export ConsulUtil as ConfigManager & refactor hax consul interface
Problem:
Consul utils are placed in common utils module hax directly interacts
with consul utils without any config abstraction.

Solution:
1. ConfigManager was exported as helper abstraction for consul
helpers in utils module.
2. Direct usage of consul util was replaced by ConfigManager
